refactor: split image_loader.rs into submodules
Closes #342
