+++ src/image_loader/scan.rs
@@ -1,0 +1,105 @@
+//! Image path scanning and directory traversal.
+
+use crate::error::{Result, SldshowError};
+use camino::Utf8PathBuf;
+use log::warn;
+use rayon::prelude::*;
+use std::path::Path;
+
+/// Maximum directory recursion depth to prevent infinite loops
+const MAX_SCAN_DEPTH: usize = 128;
+
+pub fn scan_image_paths(
+    input_paths: &[Utf8PathBuf],
+    scan_subfolders: bool,
+) -> Result<Vec<Utf8PathBuf>> {
+    let mut paths: Vec<Utf8PathBuf> = input_paths
+        .par_iter()
+        .flat_map_iter(|path| {
+            let std_path = path.as_std_path();
+            if std_path.is_file() {
+                if is_supported_image(std_path) {
+                    vec![path.clone()].into_iter()
+                } else {
+                    vec![].into_iter()
+                }
+            } else if std_path.is_dir() {
+                match scan_directory_recursive_parallel(std_path, scan_subfolders, 0) {
+                    Ok(dir_paths) => dir_paths.into_iter(),
+                    Err(e) => {
+                        warn!("Failed to scan directory {}: {}", path, e);
+                        vec![].into_iter()
+                    }
+                }
+            } else {
+                vec![].into_iter()
+            }
+        })
+        .collect();
+
+    paths.sort_by(|a, b| alphanumeric_sort::compare_str(a.as_str(), b.as_str()));
+
+    if paths.is_empty() {
+        return Err(SldshowError::NoImagesFound {
+            paths: input_paths.to_vec(),
+        });
+    }
+
+    Ok(paths)
+}
+
+fn scan_directory_recursive_parallel(
+    dir: &Path,
+    recursive: bool,
+    depth: usize,
+) -> Result<Vec<Utf8PathBuf>> {
+    if depth >= MAX_SCAN_DEPTH {
+        warn!(
+            "Maximum scan depth ({}) reached at: {}",
+            MAX_SCAN_DEPTH,
+            dir.display()
+        );
+        return Ok(Vec::new());
+    }
+
+    let entries = match std::fs::read_dir(dir) {
+        Ok(e) => e,
+        Err(e) => {
+            warn!("Failed to read directory {}: {}", dir.display(), e);
+            return Ok(Vec::new());
+        }
+    };
+
+    let paths: Vec<Utf8PathBuf> = entries
+        .flatten()
+        .par_bridge()
+        .flat_map_iter(|entry| {
+            let path = entry.path();
+            if path.is_file() && is_supported_image(&path) {
+                match Utf8PathBuf::try_from(path.clone()) {
+                    Ok(utf8_path) => vec![utf8_path].into_iter(),
+                    Err(_) => {
+                        warn!("skipping non-UTF-8 path: {:?}", path);
+                        vec![].into_iter()
+                    }
+                }
+            } else if path.is_dir() && recursive {
+                match scan_directory_recursive_parallel(&path, recursive, depth + 1) {
+                    Ok(subdir_paths) => subdir_paths.into_iter(),
+                    Err(e) => {
+                        warn!("failed to scan directory {:?}: {}", path, e);
+                        vec![].into_iter()
+                    }
+                }
+            } else {
+                vec![].into_iter()
+            }
+        })
+        .collect();
+
+    Ok(paths)
+}
+
+fn is_supported_image(path: &Path) -> bool {
+    image::ImageFormat::from_path(path).is_ok()
+}
